+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased] - ReleaseDate
 
+## [0.5.2] - 2026-05-26
+
 ### Fixed
 
 - 🐛 parsing `mark` fields in Zvuk DTOs
@@ -126,7 +128,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 🎉 Initial release
 
 <!-- next-url -->
-[Unreleased]: https://github.com/skarrok/zvuk-dl-rs/compare/v0.5.1...HEAD
+[Unreleased]: https://github.com/skarrok/zvuk-dl-rs/compare/v0.5.2...HEAD
+[0.5.2]: https://github.com/skarrok/zvuk-dl-rs/compare/v0.5.1...v0.5.2
 [0.5.1]: https://github.com/skarrok/zvuk-dl-rs/compare/v0.5.0...v0.5.1
 [0.5.0]: https://github.com/skarrok/zvuk-dl-rs/compare/v0.4.3...v0.5.0
 [0.4.3]: https://github.com/skarrok/zvuk-dl-rs/compare/v0.4.2...v0.4.3